+++ texk/dvipdfm-x/cidtype2.h
@@ -25,7 +25,7 @@
 #include "cid_p.h"
 
 extern void CIDFont_type2_set_verbose (void);
-extern void CIDFont_type2_set_flags   (long flags);
+extern void CIDFont_type2_set_flags   (int flags);
 
 extern int  CIDFont_type2_open    (CIDFont *font, const char *name,
 				   CIDSysInfo *cmap_csi, cid_opt *opt);
