+++ texk/dvipdfm-x/cs_type2.h
@@ -35,9 +35,9 @@ typedef struct {
   } seac;   /* unused in Type 2 charstring */
 } cs_ginfo;
 
-extern long cs_copy_charstring (card8 *dest, long destlen,
-				card8 *src, long srclen,
-				cff_index *gsubr, cff_index *subr,
-				double default_width, double nominal_width, cs_ginfo *ginfo);
+extern int cs_copy_charstring (card8 *dest, int destlen,
+			       card8 *src, int srclen,
+			       cff_index *gsubr, cff_index *subr,
+			       double default_width, double nominal_width, cs_ginfo *ginfo);
 
 #endif /* _CS_TYPE2_H_ */
